frontend-service/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
python-dotenv==0.19.0
itsdangerous==2.0.1
click==8.0.1